# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API - Sistema de Gestão de Ordens de Serviço",
    "version": "1.0.0",
    "description": "API REST para gerenciamento de ordens de serviço de assistência técnica.\n\n**Autores:** Diogo Pinho, Yan Marra, Gabriel Vargas"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Servidor principal"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Client": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "complement": {
            "type": "string",
            "nullable": true
          },
          "neighborhood": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ClientList": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "example": 42
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Client"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/service-orders/{id}/parts": {
      "post": {
        "tags": [
          "ServiceOrders"
        ],
        "summary": "Vincular peça à ordem (desconta estoque)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "partId",
                  "quantity"
                ],
                "properties": {
                  "partId": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Peça vinculada com sucesso"
          },
          "422": {
            "description": "Peça sem estoque ou estoque insuficiente"
          }
        }
      }
    },
...
    "/parts": {
      "get": {
        "tags": [
          "Parts"
        ],
        "summary": "Listar todas as peças",
        "responses": {
          "200": {
            "description": "Lista de peças com estoque"
          }
        }
      },
      "post": {
        "tags": [
          "Parts"
        ],
        "summary": "Cadastrar nova peça (somente admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "code",
                  "price"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Tela LCD 6.5\""
                  },
                  "code": {
                    "type": "string",
                    "example": "TELA-LCD-65"
                  },
                  "description": {
                    "type": "string"
                  },
                  "stock": {
                    "type": "integer",
                    "default": 0
                  },
                  "unit": {
                    "type": "string",
                    "default": "un"
                  },
                  "price": {
                    "type": "number",
                    "example": 180
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Peça criada"
          },
          "409": {
            "description": "Código já cadastrado"
          }
        }
      }
    },
    "/parts/{id}": {
      "get": {
        "tags": [
          "Parts"
        ],
        "summary": "Buscar peça por ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados da peça"
          },
          "404": {
            "description": "Não encontrada"
          }
        }
      },
      "put": {
        "tags": [
          "Parts"
        ],
        "summary": "Atualizar peça (somente admin)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Peça atualizada"
          }
        }
      },
      "delete": {
        "tags": [
          "Parts"
        ],
        "summary": "Remover peça (somente admin)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removida com sucesso"
          }
        }
      }
    },
    "/parts/{id}/restock": {
      "post": {
        "tags": [
          "Parts"
        ],
        "summary": "Repor estoque de uma peça (gera histórico de movimentação)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "quantity"
                ],
                "properties": {
                  "quantity": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estoque atualizado"
          }
        }
      }
    },
    "/parts/{id}/movements": {
      "get": {
        "tags": [
          "Parts"
        ],
        "summary": "Histórico de movimentações de estoque de uma peça",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico de entrada e saída"
          }
        }
      }
    },
    "/clients/search": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Buscar clientes por nome, CPF, e-mail ou telefone",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Termo de busca"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de clientes encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/clients": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Listar todos os clientes (paginado)",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de clientes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Cadastrar novo cliente (CEP preenchido automaticamente via ViaCEP)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "cpf",
                  "phone"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Maria Silva"
                  },
                  "cpf": {
                    "type": "string",
                    "example": "123.456.789-00"
                  },
                  "phone": {
                    "type": "string",
                    "example": "(11) 99999-9999"
                  },
                  "email": {
                    "type": "string",
                    "example": "maria@email.com"
                  },
                  "zipCode": {
                    "type": "string",
                    "example": "01310-100"
                  },
                  "number": {
                    "type": "string",
                    "example": 42
                  },
                  "complement": {
                    "type": "string",
                    "example": "Apto 3"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          },
          "409": {
            "description": "CPF já cadastrado"
          }
        }
      }
    },
    "/clients/{id}": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Buscar cliente por ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do cliente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Clients"
        ],
        "summary": "Atualizar cliente",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "zipCode": {
                    "type": "string"
                  },
                  "number": {
                    "type": "string"
                  },
                  "complement": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cliente atualizado"
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Clients"
        ],
        "summary": "Remover cliente",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removido com sucesso"
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        }
      }
    },
...
  },
...
}